Extract base class / move Coordinators::ControllerCoordinator
