+++ lib/filterameter/declarative_controller_filters.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'filterameter/filterable'
-require 'filterameter/controller_filters'
+require 'filterameter/coordinators/controller_coordinator'
 
 module Filterameter
   # = Declarative Controller Filters
@@ -13,26 +13,26 @@ module Filterameter
 
     class_methods do
       def filter_model(model_class, query_var_name = nil)
-        controller_filters.model_class = model_class
+        filter_coordinator.model_class = model_class
         filter_query_var_name(query_var_name) if query_var_name.present?
       end
 
       def filter_query_var_name(query_variable_name)
-        controller_filters.query_variable_name = query_variable_name
+        filter_coordinator.query_variable_name = query_variable_name
       end
 
-      def controller_filters
-        @controller_filters ||= Filterameter::ControllerFilters.new(controller_name, controller_path)
+      def filter_coordinator
+        @filter_coordinator ||= Filterameter::Coordinators::ControllerCoordinator.new(controller_name, controller_path)
       end
     end
 
     private
 
     def build_filtered_query
-      var_name = "@#{self.class.controller_filters.query_variable_name}"
+      var_name = "@#{self.class.filter_coordinator.query_variable_name}"
       instance_variable_set(
         var_name,
-        self.class.controller_filters.build_query(params.to_unsafe_h.fetch(:filter, {}),
+        self.class.filter_coordinator.build_query(params.to_unsafe_h.fetch(:filter, {}),
                                                   instance_variable_get(var_name))
       )
     end
